requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytest-django==4.5.2
pytest-mock==3.10.0
pytz==2022.7.1
sqlparse==0.4.3
termcolor==2.2.0
tomli==2.0.1
